Fix nav bar visible-on-ancestor for the Polymer to Lit dropdown conversion
In the nav bar, the visible-on-ancestor functionality broke because the Lit version loads
faster than the Polymer version. This means that when the visible-on-ancestor-mixin
uses findComposedAncestor, the parent element has loaded but its shadow DOM has not.
Because the 'd2l-visible-on-ancestor-target' is hidden within the ancestor's shadow
DOM, the look up fails. To fix this, 'd2l-visible-on-ancestor-target' needs to be
moved outside of the ancestor's shadow DOM into the light DOM.