Support a new `widthType` property.
The `widthType` supports `normal` and `fullscreen`.
`normal` restricts the nav bar left/middle/right to fit a max-width of `1230px`.
`fullscreen` sets the max-width to be '100%'

Note that there is existing support for a css variable that sets the maxWidth of this same element.
`max-width: var(--d2l-navigation-immersive-content-max-width, 100%);`

It is possible for there to be a clash if the user sets both the css variable and sets this `widthType` property. If the user does not set either, the `max-width` will be `100%`