Update README.md
There doesn't appear to be a clear command to install this component, similar to what is avialable on https://github.com/BrightspaceUI/core.  Trying an NPM install of the same format that's available for CORE components doesn't work here so I've added the NPM install command that works for this navigation component.